requirements.txt
django
django-hvad
django-countries
django-libs
